Aura/services/auth-service/pyproject.toml
[tool.poetry]
name = "aura-auth-service"
version = "0.1.0"
description = "Authentication service for the Aura platform"
authors = ["Aura Team <info@aura.ai>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
email-validator = "^2.0.0"
httpx = "^0.24.0"
aura-common = {path = "../../libs/common", develop = true}

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 